build a binary-only alpine docker image
Previous docker image included the entire debian build image, which was
huge. With this change we migrate to alpine linux, and we separate the
build image from the cflakes binary image itself. The build image is
still pretty big, but since the main binary image has no runtime
dependencies, it's tiny.